refactor: 패키지 line, station, interstation 으로 구분
