+++ src/main/java/subway/controller/dto/InterStationResponse.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import subway.domain.InterStation;
+import subway.domain.interstation.InterStation;
 
 @Getter
 @AllArgsConstructor
