+++ src/main/java/subway/repository/LineRepositoryImpl.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import subway.dao.InterStationDao;
 import subway.dao.LineDao;
 import subway.dao.StationDao;
-import subway.domain.InterStation;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.interstation.InterStation;
+import subway.domain.line.Line;
+import subway.domain.station.Station;
 import subway.entity.InterStationEntity;
 import subway.entity.LineEntity;
 import subway.entity.StationEntity;
